+++ common/cpp/src/flutter_webrtc.cc
@@ -19,7 +19,12 @@ FlutterWebRTC::~FlutterWebRTC() {}
 void FlutterWebRTC::HandleMethodCall(
     const MethodCallProxy& method_call,
     std::unique_ptr<MethodResultProxy> result) {
-  if (method_call.method_name().compare("createPeerConnection") == 0) {
+  if (method_call.method_name().compare("initialize") == 0) {
+   const EncodableMap params =
+       GetValue<EncodableMap>(*method_call.arguments());
+   const EncodableMap options = findMap(params, "options");
+   result->Success();
+  } else if (method_call.method_name().compare("createPeerConnection") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null arguments received");
       return;
@@ -425,8 +430,7 @@ void FlutterWebRTC::HandleMethodCall(
     const std::string peerConnectionId = findString(params, "peerConnectionId");
     RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
     if (pc == nullptr) {
-      result->Error("peerConnectionDisposeFailed",
-                    "peerConnectionDisposeClose() peerConnection is null");
+      result->Success();
       return;
     }
     RTCPeerConnectionDispose(pc, peerConnectionId, std::move(result));