+++ lib/flutter_webrtc.dart
@@ -7,7 +7,6 @@ export 'src/helper.dart';
 export 'src/desktop_capturer.dart';
 export 'src/media_devices.dart';
 export 'src/media_recorder.dart';
-export 'src/frame_cryptor.dart';
 export 'src/native/factory_impl.dart'
     if (dart.library.html) 'src/web/factory_impl.dart';
 export 'src/native/motion_detection/detection_frame.dart';
@@ -20,3 +19,6 @@ export 'src/native/rtc_video_view_impl.dart'
 export 'src/native/utils.dart' if (dart.library.html) 'src/web/utils.dart';
 export 'src/native/video_recorder/rtc_record_result.dart';
 export 'src/native/video_recorder/video_recorder.dart';
+export 'src/native/adapter_type.dart';
+export 'src/native/android/audio_configuration.dart';
+export 'src/native/ios/audio_configuration.dart';