+++ lib/src/native/media_recorder_impl.dart
@@ -3,8 +3,7 @@ import 'dart:math';
 
 import 'package:webrtc_interface/webrtc_interface.dart';
 
-import '../native/media_stream_track_impl.dart';
-
+import 'media_stream_track_impl.dart';
 import 'utils.dart';
 
 class MediaRecorderNative extends MediaRecorder {
@@ -12,15 +11,10 @@ class MediaRecorderNative extends MediaRecorder {
   final _recorderId = _random.nextInt(0x7FFFFFFF);
 
   @override
-  Future<void> start(
-    String path, {
-    MediaStreamTrack? videoTrack,
-    RecorderAudioChannel? audioChannel,
-    int? videoWidth,
-    int? videoHeight,
-
-    // TODO(cloudwebrtc): add codec/quality options
-  }) async {
+  Future<void> start(String path,
+      {MediaStreamTrack? videoTrack, RecorderAudioChannel? audioChannel
+      // TODO(cloudwebrtc): add codec/quality options
+      }) async {
     if (audioChannel == null && videoTrack == null) {
       throw Exception('Neither audio nor video track were provided');
     }
@@ -30,9 +24,9 @@ class MediaRecorderNative extends MediaRecorder {
       if (audioChannel != null) 'audioChannel': audioChannel.index,
       if (videoTrack != null) 'videoTrackId': videoTrack.id,
       'recorderId': _recorderId,
-      'peerConnectionId': videoTrack is MediaStreamTrackNative ? videoTrack.peerConnectionId : null
-      'videoWidth': videoWidth,
-      'videoHeight': videoHeight,
+      'peerConnectionId': videoTrack is MediaStreamTrackNative
+          ? videoTrack.peerConnectionId
+          : null
     });
   }